Refactor properties into separate files
