+++ tests/test_prop_event.py
@@ -7,7 +7,8 @@ import pytest
 
 from custom_components.yandex_smart_home import const
 from custom_components.yandex_smart_home.error import SmartHomeError
-from custom_components.yandex_smart_home.prop import PROPERTY_EVENT, PROPERTY_FLOAT, _EventProperty
+from custom_components.yandex_smart_home.prop_event import PROPERTY_EVENT, _EventProperty
+from custom_components.yandex_smart_home.prop_float import PROPERTY_FLOAT
 
 from . import BASIC_CONFIG
 from .test_prop import assert_no_properties, get_exact_one_property
